Performance improvements and refactoring
## The `RDFWriter.getResource() method

The method was used to avoid emitting a `X rdf:typ Y` triple *more than once* (i.e. each time X was encountered in the IFC). However, the code was written in such a way that many times, the X that was checked if it had been used before, was actually generated to be unique, and the check was unnecessary. In the rest of the cases, it was just a convenience method used to make sure that X was emitted *at least once*. Analysis showed that this is not necessary. The change allows us to remove the whole `resources` map, which is a big memory and performance improvement.

## Resource naming in `RDFWriter`

Naming the resources used to depend on the global `idCounter` variable in RDFWriter, which was incremented after each use. This construction made it impossible to use parallelization as the order of resource creation is not deterministic in that case. Therefore, a deterministic approach was chosen for naming the resources such that the probability of a name clash is small and the name is always the same for the same logical resource. This is done using the method `makeResourceWithHashedSuffix` (and its helpers), which generates a cryptographic hash of some 'key components' provided by the caller and appends the hash to the resource's 'base name'. The combination of these components being unique, the resulting URI is unique. This change required the tests to be re-generated

## Avoid Duplicating Property Resources made optional

The `RDFWriter` used to generate one property/value resource for all instances of the property value in the IFC. This is now optional, depending on the property `RDFWriter.avoidDuplicatePropertyResources`. Not doing it produces bigger RDF models but is considerably faster and requires less memory. The tests for this option have been added and the expected output generated.

## Parallelized RDF generation

With all the above changes it became possible to parallelize the generation of RDF. The extracted, thread-safe method `RDFWriter.generateTriplesForIfcVo(IFCVO)` is the central point for that.

## Renaming of the test folders

In order for tests to be easily understandable, the folder `converIFCFileToOutputTTL` was renamed to `convert` and another folder `convert_avoidDuplicatePropertyResources` was created for the new tests. These names directly map to the method names in the `TestIfcSpfReader` class.